flask_app/requirements.txt
tornado
Pillow
flask
pymongo